Fixed Hugging Face API calls
Sorry for the late edit, but I have been periodically checking on this Streamlit app (because this assignment has not been graded after many weeks of being due). Today, I found that the app would not load anymore, so I had to debug the issue. I think I was rate-limited by the Hugging Face API. Hopefully, this fixes that issue.